update native coins
