+++ vm/src/nova_stdlib/sources/native_coins.move
@@ -1,0 +1,3 @@
+module nova_std::native_coins {
+    struct Nova {}
+}
